requirements.txt
# Core database clients
mysql-connector-python>=8.1.0
psycopg2-binary>=2.9.7

# AWS SDK
boto3>=1.26.0
botocore>=1.29.0

# Data validation
pydantic>=2.3.0
PyYAML>=6.0

# CLI and utility
argparse>=1.4.0
python-dotenv>=1.0.0

# Testing / Development
pytest>=7.4.0
pytest-mock>=3.12.0
mock>=5.0.1

# Optional: For file compression
python-magic>=0.4.28